# swagger/spring-server-generated/src/main/java/io/swagger/model/DecryptionResponse.java
package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.v3.oas.annotations.media.Schema;
import org.springframework.validation.annotation.Validated;
import javax.validation.Valid;
import javax.validation.constraints.*;

/**
 * DecryptionResponse
 */
@Validated
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-10-10T17:13:44.864504617Z[GMT]")


public class DecryptionResponse   {
  @JsonProperty("decryptedMessage")
  private String decryptedMessage = null;

  public DecryptionResponse decryptedMessage(String decryptedMessage) {
    this.decryptedMessage = decryptedMessage;
    return this;
  }

  /**
   * Get decryptedMessage
   * @return decryptedMessage
   **/
  @Schema(example = "4512", required = true, description = "")
      @NotNull

    public String getDecryptedMessage() {
    return decryptedMessage;
  }

  public void setDecryptedMessage(String decryptedMessage) {
    this.decryptedMessage = decryptedMessage;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    DecryptionResponse decryptionResponse = (DecryptionResponse) o;
    return Objects.equals(this.decryptedMessage, decryptionResponse.decryptedMessage);
  }

  @Override
  public int hashCode() {
    return Objects.hash(decryptedMessage);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class DecryptionResponse {\n");
    
    sb.append("    decryptedMessage: ").append(toIndentedString(decryptedMessage)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}